[refactor] VO의 위치 변경
