+++ reservation-service/src/main/java/com/rainbowgon/reservationservice/global/vo/TimeVO.java
@@ -1,4 +1,4 @@
-package com.rainbowgon.reservationservice.domain.reservation.vo;
+package com.rainbowgon.reservationservice.global.vo;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.rainbowgon.reservationservice.domain.reservation.entity.AvailableStatus;
